+++ packages/navi-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @awell-health/navi-core
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated auth service to use int array instead of crypto
+
 ## 0.0.1
 
 ### Patch Changes