+++ packages/navi-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @awell-health/navi-js-react
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated auth service to use int array instead of crypto
+- Updated dependencies
+  - @awell-health/navi-core@0.0.2
+  - @awell-health/navi-js@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes